Style: Refactor Register Info Dialog in InfoFormPage 등록
