+++ android/app/src/main/AndroidManifest.xml
@@ -4,6 +4,10 @@
         android:label="iris_flutter"
         android:name="${applicationName}"
         android:icon="@mipmap/ic_launcher">
+
+        <meta-data android:name="com.google.android.geo.API_KEY"
+            android:value="AIzaSyA9SFLk4JB_fWuLIeEOwmylOgfOcISY--g"/>
+
         <activity
             android:name=".MainActivity"
             android:exported="true"
