+++ pubspec.yaml
@@ -36,10 +36,10 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   get: ^4.6.6
-  google_maps_flutter: ^2.5.3
-  carousel_slider: ^4.2.1
   image_picker: ^1.0.7
   smooth_page_indicator: ^1.1.0
+  carousel_slider: ^4.2.1
+  google_maps_flutter: ^2.5.3
 
 dev_dependencies:
   flutter_test:
